[#8] chore: AppLayout 폴더 구조 변경
